requirement.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djongo==1.3.6
Flask==2.2.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==3.20.1
pymongo==3.12.0
pytz==2022.1
SQLAlchemy==1.4.40
sqlparse==0.2.4
tzdata==2022.1
Werkzeug==2.2.1
WTForms==3.0.1
